Update to README
Suggest changing SKL to SKALE.

There is already confusion around SKL and sFuel on SKALE chains.

With SKALE G on the way with a non-zero gas fee and people used to paying for Oracle services on other chains, referencing SKL in this doc may lead to a SKALE chain user inferring they need to pay SKL to use SKALE Oracle.